refactor - 패키지 구조 수정 및 엔드포인트 수정
