+++ app/src/main/java/com/example/sport_planet/data/response/MakeChattingRoomResponse.kt
@@ -1,4 +1,4 @@
-package com.example.sport_planet.model
+package com.example.sport_planet.data.response
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
