+++ app/src/main/java/com/example/sport_planet/presentation/chatting/ChattingConstant.kt
@@ -4,5 +4,5 @@ object ChattingConstant {
     val CHAT_BOT_TYPE = "BOT_MESSAGE"
     val PROFILE_TYPE = "PROFILE"
     val TALK_TYPE = "TALK"
-    val URL = "ws://54.180.29.231/api/chatting-service/ws/chat/websocket"
+    val URL = "ws://54.180.29.231:8080/ws/chat/websocket"
 }
