+++ app/src/main/java/com/example/sport_planet/presentation/chatting/ChattingFragment.kt
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.sport_planet.R
 import com.example.sport_planet.data.enums.SeparatorEnum
 import com.example.sport_planet.databinding.FragmentChattingBinding
-import com.example.sport_planet.model.ChattingRoomListResponse
+import com.example.sport_planet.data.response.ChattingRoomListResponse
 import com.example.sport_planet.presentation.base.BaseFragment
 import com.example.sport_planet.presentation.base.BaseViewModel
 import com.example.sport_planet.presentation.chatting.adapter.ChattingRoomAdapter
