+++ app/src/main/java/com/example/sport_planet/remote/Api.kt
@@ -6,10 +6,7 @@ import com.example.sport_planet.data.request.MyViewEditRequest
 import com.example.sport_planet.data.response.*
 import io.reactivex.Single
 import retrofit2.http.*
-import com.example.sport_planet.model.*
 import com.beust.klaxon.JsonObject
-import io.reactivex.Single
-import retrofit2.http.*
 const val jwt = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJ0ZXN0ZXIxIiwiaWF0IjoxNjA0NjY3MzA5LCJleHAiOjE2MzYyMDMzMTEsImF1ZCI6IiIsInN1YiI6InRlc3RlcjFAZ21haWwuY29tIiwidXNlcklkIjoiMSJ9.Bmbhc-I1r-L-dW5vUzvB9jRsPPKtcqYXutAyWKqkPrc"
 //const val jwt = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJ0ZXN0ZXIyIiwiaWF0IjoxNjA0NjY3MzA5LCJleHAiOjE2MzYyMDMzMTEsImF1ZCI6IiIsInN1YiI6InRlc3RlcjJAZ21haWwuY29tIiwidXNlcklkIjoiMiJ9.iMoSPe9k5Uj6w8qa4eaOIQqjIzI5Scuts3tBbQ4p79g"
 
@@ -25,7 +22,6 @@ interface Api {
     fun postSignIn(@Body userInfo: LoginResponse): Single<ServerCallBackResponse>
 
     @POST("/v1/user/sign-up")
-
     fun postSignUp(@Body userSignUp: SignUpResponse): Single<ServerCallBackResponse>
 
     // 마이페이지 API
@@ -68,8 +64,6 @@ interface Api {
 
     @POST("/v1/user/my-profile/history/evaluate/report")
     fun postEvaluateReport(@Body evaluateReportRequest: EvaluateReportRequest): Single<ServerCallBackResponse>
-
-    fun postSignUp(@Body userSignUp : SignUpResponse) : Single<ServerCallBackResponse>
 
     @Headers("Authorization: Bearer $jwt")
     @POST("/api/chatting-service/v1/chat/room")
