+++ app/src/main/res/layout/item_notice_message.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="itemNoticeMessage"
-            type="com.example.sport_planet.model.ChattingMessageResponse" />
+            type="com.example.sport_planet.data.response.ChattingMessageResponse" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
