+++ app/src/main/res/layout/item_received_talk_message.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="ItemReceivedTalkMessage"
-            type="com.example.sport_planet.model.ChattingMessageResponse" />
+            type="com.example.sport_planet.data.response.ChattingMessageResponse" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
