+++ app/src/main/res/layout/item_sent_profile_message.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="ItemSentProfileMessage"
-            type="com.example.sport_planet.model.ChattingMessageResponse" />
+            type="com.example.sport_planet.data.response.ChattingMessageResponse" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
